refactor: organiza pacotes de interfaces e implementações
- Move as interfaces `RegisterUserUseCase` e `UpdateUserUseCase` para o pacote `ports.api` para melhor separação de responsabilidades.
- Atualiza as classes e `Controller` afetados para refletir a nova estrutura de pacotes.

RefS: GH-1
