+++ src/main/java/dev/ghonda/project/management/users/ports/api/UpdateUserUseCase.java
@@ -1,4 +1,4 @@
-package dev.ghonda.project.management.users.usecases;
+package dev.ghonda.project.management.users.ports.api;
 
 import dev.ghonda.project.management.users.rest.dto.UpdateUserPayload;
