+++ src/main/java/dev/ghonda/project/management/users/rest/UserController.java
@@ -4,8 +4,8 @@ import dev.ghonda.project.management.shared.dto.ApiResponse;
 import dev.ghonda.project.management.shared.dto.Resource;
 import dev.ghonda.project.management.users.rest.dto.RegisterUserPayload;
 import dev.ghonda.project.management.users.rest.dto.UpdateUserPayload;
-import dev.ghonda.project.management.users.usecases.RegisterUserUseCase;
-import dev.ghonda.project.management.users.usecases.UpdateUserUseCase;
+import dev.ghonda.project.management.users.ports.api.RegisterUserUseCase;
+import dev.ghonda.project.management.users.ports.api.UpdateUserUseCase;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
