+++ src/main/java/dev/ghonda/project/management/users/usecases/RegisterUserUseCaseImpl.java
@@ -7,6 +7,7 @@ import dev.ghonda.project.management.shared.validators.ValidatorService;
 import dev.ghonda.project.management.users.domain.Role;
 import dev.ghonda.project.management.users.domain.User;
 import dev.ghonda.project.management.users.domain.UserJpaRepository;
+import dev.ghonda.project.management.users.ports.api.RegisterUserUseCase;
 import dev.ghonda.project.management.users.rest.dto.RegisterUserPayload;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
