+++ src/main/java/dev/ghonda/project/management/users/usecases/UpdateUserUseCaseImpl.java
@@ -5,6 +5,7 @@ import dev.ghonda.project.management.shared.exceptions.ResourceNotFoundException
 import dev.ghonda.project.management.shared.validators.ValidatorService;
 import dev.ghonda.project.management.users.domain.Role;
 import dev.ghonda.project.management.users.domain.UserJpaRepository;
+import dev.ghonda.project.management.users.ports.api.UpdateUserUseCase;
 import dev.ghonda.project.management.users.rest.dto.UpdateUserPayload;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
